pygr: Refactor CoordConverter / reduced functionality (API change)
In order to allow conversions between normalized device coordinates (NDC)
and world coordinates (WC) independet of width and height attributes remove
device coordinates (DC) conversion from CoordConverter and the
corresponding width and height attributes. This functionality can still be
used with the new derived class DeviceCoordConverter. Unfortunately the API
for CoordConverter will be changed by this patch but it is conventional to
call the base class CoordConverter and the derived class
DeviceCoordConverter and not vice versa.

For existing programs just replace CoordConverter with DeviceCoordConverter.

Change-Id: I94f5a3b2b5defce7be9a043df639b7aa38258618
Reviewed-on: https://iffwww.iff.kfa-juelich.de/review/55
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>