pygr: New Text object for drawing text on axes or in ndc.
If a PlotAxes object has been passed to the Text object the
text will be drawn in world coordinates x, y on the axes.
If the bounding box of the text is out of bounds of corresponding
viewport the text will not be drawn by default. This can be changed
by setting hideviewport to False.

Change-Id: I1766500657177dd627093ce239ff286972f25689
Reviewed-on: https://iffwww.iff.kfa-juelich.de/review/56
Reviewed-by: Josef Heinen <j.heinen@fz-juelich.de>