pygr: PlotAxes: Fix unexpected reset of window.
_resetWindow has been set to True on default to calculate a window
on the first drawGR call. If a window has been specified before the
first drawGR call has overwritten the window. Therefore set _resetWindow
to False on default and check for window is None inReset. This ensures
that a window will be calculated if no window has been specified but
does not overwrite an existing one.

Change-Id: Ifdb5856956dd285091e65c6fc1fa5a7f078ef5ee
Reviewed-on: https://iffwww.iff.kfa-juelich.de/review/82
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>
Reviewed-by: Josef Heinen <j.heinen@fz-juelich.de>