pygr: PlotAxes: Make viewport mandatory (API change).
The viewport attribute has been made mandatory to avoid different
viewport settings between Plot and PlotAxes objects. Even though the Plot
object sets the viewport on the PlotAxes object when the axes object has
been added with addAxes this can lead to wrong results, e.g. in the
following case. If there are calculations relying on the current viewport
in the PlotAxes object before this has been added to a Plot object and
no default viewport has been used the viewport for the Plot and the PlotAxes
object differs.

Change-Id: I66f025bba3939e07912d2fd239835e5e11eb5b2b
Reviewed-on: https://iffwww.iff.kfa-juelich.de/review/85
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>