Fix: Setting wrong display list length on fctid 6 (clear workstation).
The display list buffer contains the length for each display list item
as first element (stored as integer). The end of the display list will
be marked by setting the length to zero. On clear workstation (fctid 6) the
display list buffer will be cleared or rather the first entry will be set
to zero. In several plugins the buffer has been interpreted as char array
because of its declaration as char * but the program logic expected
interpreting this buffer as byte array which corresponds to char * in C
programming language (1 char = 1 Byte). Thus setting the first item of the
buffer to '\0' just sets the first Byte to 0 and leads to length 1024 for
the first display list item. At buffer address 1024 on 64 bit the value was
accidentally 0 whereas on 32 bit platforms the value was some unreasonable
value for addressing the buffer (see debugging output below). This resulted
to an SEGV on 32 bit platforms. In order to fix this issue the first
sizeof(int) Bytes will be set to 0 in this patch.

64bit:
[jcns@centos-6.shared gr (setup-qhull)]$ python examples/qtgrdemo.py
len pointer: 0x7f26c23a2010
      value: 1024
[w] len pointer: 0x7f26c23a2410
[w]       value: 0
end while

32bit:
[jcns@centos-6-i686.shared gr (setup-qhull)]$ python examples/qtgrdemo.py
len pointer: 0xb67c1008
      value: 1024
[w] len pointer: 0xb67c1408
[w]       value: 1717986918
end while

Change-Id: Ifa6e93b15bb50a46640aa2048ac582d6adfbfabb
Reviewed-on: https://iffwww.iff.kfa-juelich.de/review/608
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>