Fixed Qt version detection for PyQt5 applications on Linux
For Qt applications written in Python the gks plugin loader was not able
to determine the running Qt version on Linux and loaded the Qt4 plugin
as a fallback (resulting in a segmentation fault).
As a workaround, the qtgr backend now loads the QtCore wrapper (which is
a c extension module) twice via ctypes CDLL and exports all included
symbols globally (for example the qVersion function) and the gks plugin
loader can dissolve the qVersion symbol again.

Change-Id: I28097c49d5412cb013d0bd5cdc4b28bcc75fda9b
Reviewed-on: https://iffwww.iff.kfa-juelich.de/review/731
Reviewed-by: Josef Heinen <j.heinen@fz-juelich.de>