setup.py: qmake executables without version suffix are preferred
If the qmake executables is located automatically, qmake executables
without version suffix are preferred now (but the qt version is also
checked). Preferring qmake without suffix is important for conda
builds since conda qt only has a qmake executable without version
suffix. If qmake with suffix would be preferred the system qmake
could be chosen by mistake.

The setup.py script is now able to detect if Qt5 on macOS was built
as a framework. That is also important for building conda packages
because conda ships with Qt libraries that are not built as a
macOS framework.

Bugfix: If Qt4 qmake executables was specified, it was also taken
for Qt5.

Added support for Qt4 framework builds

Change-Id: Ia3eb584e7b9b41557fe6d7a8ad344c9803b3b646
Reviewed-on: https://iffwww.iff.kfa-juelich.de/review/758
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>
Reviewed-by: Josef Heinen <j.heinen@fz-juelich.de>