qtgr: Fix world coordinate calculation for multiple plots.
For multiple plots in one widget it is necessary to pass the
current window for a `PlotAxes` object to the corresponding
`MouseEvent` in order to calculate world coordinates properly.
Otherwise the internal state machine's window will be used which
corresponds to the last issued ``gr.setwindow`` command
(last drawn `PlotAxes`).

Change-Id: I5fadcea84eb38ee8d5078a95f3e696b8bc2387ad
Reviewed-on: https://iffwww.iff.kfa-juelich.de/review/785
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>