gks: fix a Qt performance issue
Starting with GKS version 5, the linewidth is adjusted according the window
height. This leads to problems when drawing lines with a large number of
points. The line drawing speed slows down extremely on all platforms,
probably due a faulty line drawing algorithm in Qt 4 and 5.

With this patch, the linewidth is not adjusted if the number of points
is larger than half the product of the linewidth and the window height.
This approach is a simple workaround and does not produce any other artefact.
The real problem must be examined further.

Change-Id: I302275aa0ae7fd199099bebfa857e90ec6dced57
Reviewed-on: https://iffwww.iff.kfa-juelich.de/review/802
Reviewed-by: Josef Heinen <j.heinen@fz-juelich.de>