pygr, qtgr: Fix mouse events possibly return wrong x, y values.
If multiple widgets are open the last drawn will store it's
state, e.g. window and scale in the gks state machine which leads
to wrong calculations using window and scale in CoordConverter
because this is not used inside one paint event.
Ensure window and scale are set by corresponding axes and use these
values for calculation.

Change-Id: Ic88c4f5916b5a921344b816f8d9113f1ff314d9b
Reviewed-on: https://iffwww.iff.kfa-juelich.de/review/834
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>