Increase work buffer size for gridit / idsfft
The gridit code uses a work buffer to store the interpolation parameters
for each triangle in the point set's triangulation. Previously, 5 * n has
been used, however the number of triangles in a point set's triangulation
is t = 2*n-h-2, with h being the number of points on the point sets convex
hull. As the convex hull can consist of as little as three points, the code
should not rely on h being a significant fraction of n. With the addition of
four points at the minima and maxima of the point set, this leads to h >= 4
and t = 2*(n+4)-h-2 <= 2*(n+1).
For each triangle, three interpolation parameters are stored, so in this
change, the size of the work buffer is set to 3*t = 6*(n+1) instead of 5*n.

Change-Id: I80275fad44d26c4f92069d607e20d7ffbfdce402
Reviewed-on: https://iffwww.iff.kfa-juelich.de/review/843
Reviewed-by: Josef Heinen <j.heinen@fz-juelich.de>