`meta.c` can be compiled on ARM architectures
Variable argument lists in `meta.c` used the platform specific
implementation of `va_list` on x86 / x86_64 hardware. `va_list` is
typically declared as an one-element-array and thus is converted to a
pointer type when used in function signatures. However, this is not true
for ARM hardware. For this reason, passing a `NULL` pointer to a
`va_list` parameter works on x86 / x86_64 but does **NOT** work on ARM.
This patch uses `va_list *` instead of `va_list` to avoid these
problems.