Fixed GKSTerm connection bug
The quartzplugin keeps track on windows opening and closing, but previously an
existing plugin object was seen as a working connection. Quitting the GKSTerm
application leaves the quartzplugin with an existing plugin with no
connection however, so the following OPEN_WS call to re-open a window will
cause an NSException.

This commit fixes this issue by verifying that an existing plugin object has
a working connection during the OPEN_WS call and recreating it if necessary.