The stored value format of the argument container is normalized
Before this commit, the value format of the argument container was
simply copied from the user input. However, this format does not
necessarily fit to the internal data storage. Now, the value format
describes exactly which memory layout is used internally (and thus is
normalized as well). The value format can be passed along the memory
buffer to other routines like the JSON serializer.