`g` format specifier is used for double serialization
Before this commit, the `%f` format was used to convert doubles to a
string representation. Now, `%.17g` is used instead. The `g` format
switches automatically from fixed-point to exponential notation for
small and huge numbers (-> it is guaranteed that a precision of 17 is
sufficient to represent all significant digits). The value `17` is
expanded from the macro `DBL_DECIMAL_DIG` defined in `float.h`.