Use standard character strings instead of wide character strings
Before this commit, the networking code for Windows used wide character
strings for error messages. For printing these kind of strings, the
Microsoft `printf` version offers a `%S` format specifier which is not
part of the C standard. Without wide character strings, the error
handling code conforms to the C(89) standard.