Added post processing for serialized JSON string components
The JSON serializer uses the `printf` function internally for converting
primitive datatypes to strings. However, it is not possible to convert
doubles to strings with a single `printf` that cuts trailing zeros and
guarantees a decimal dot for non-exponential outputs. Numbers without a
decimal point can be interpreted as integers by the JSON parser
unintendedly. Therefore, this commit adds a post processing step which
adds a decimal point if necessary.