Move GLX-dependent code to external dynamic library
This commit adds a libGR3platform.so to Linux builds, which links against X11
libraries and libGL, so that libGR3.so can be successfully loaded on Linux
systems which lack X11 or OpenGL support. At runtime, when gr3_init is called,
it will fail however when the required libraries are missing. This is so that
code that links against GR3 can run even on headless systems where it can then
use gr3_init to detect whether or not GR3 can run on the system. This way a
fallback can be used instead of simply being unable to run the executable.

Windows and macOS could use the same system to avoid linking against OpenGL
and window system libraries, however these systems normally provide all
required libraries.

The libGR3platform.so will first be searched in GRDIR/lib, then fall back to
regular shared library search.