Allow the use of custom send and receive function in `meta.c`
This commit adds two optional arguments to `gr_openmeta` to replace the
default send and/or receive functions for socket based communication.
These arguments are intended to be used in Jupyter kernels to allow the
use of Jupyter communicators.

`gr_openmeta` works with more abstract arguments. Instead of a hostname
and port number the routine now expects a more general string descriptor
and an id which are passed to the callback functions. These arguments
can be used to identify opened communication channels (their exact
meaning can be defined by the api user). In case of the standard socket
communication, the string descriptor is the hostname and the id the port
number, as before.

The receive callback must return a full dataset as null-terminated
string. After returning the memory for the string can be released since
it is copied internally. The send function gets a null-terminated string
when called.