Improved timeout logic in quartzplugin
When GKSTerm is in the background, its response time can vary wildly and it
can exceed the current 50ms timeout. The changes in this commit reduce the
time between checking for a response from 10ms to 1ms and increase the timeout
for the GKSTerm 'is running' response from 50ms to 500ms after the first
succesful 'is running' check.