Ignore X errors during Pbuffer creation
GR3 tries every framebuffer configuration which the system reports as suitable
for a RGBA Pbuffer, until one is found that actually works. On some systems,
e.g. in a docker container with CentOS 7 and Xvfb, there will be framebuffer
configurations which not only don't work for creating a Pbuffer, but cause a
BadValue error when GLX tries to create an internal Pixmap using the given
configuration's bit depth. If uncaught, these errors abort the process.

This commit adds an X error handler solely for the Pbuffer creation, which
ignores any X errors. This way, the broken configurations will fail quietly
and GR3 will continue until it finds a working configuration or all
configurations have failed.