Add a generic list container type
Add a generic list datatype that can be used as
- list
- queue
- stack
Stored items will be copied when added and deleted when removed from the
list.

Additionally, add a derived reference list type which only holds
references to stored objects (objects won't be copied and won't be
deleted in this list type if not explicitly requested).

This new data types replace the stack implementation and will provide a
queue for adding event handling to the meta layer.