[Meta] Introduce a new libGRM library for the meta layer
This commit adds a new libGRM library to avoid cyclic dependencies
between libGR and libGR3 (the meta layer needs both). This is the first
step of refactoring. All meta related code is put into a new library
directory `lib/grm` as a single source file with a separate header file
`grm.h`. The code is compiled into a static and a shared library
`libGRM.(a|so|dylib)` which depends on libGR and libGR3. Next commits
will split the source code into logical translation units.