[GRM] Implement type checks and automatic conversions
This MR adds a type check to GRM. Passed values can be rejected if their
type is not listed in `key_to_formats`. If a type does not match but is
compatible (e.g. a single value is given but an array expected), the
type will be converted automatically.

This commit is needed for MR !428.