[GRM] Restructure GRM tests
This commit restructures GRM tests to allow to write tests for internal
functions. In order to achieve this, a custom CMakeLists is added to
build GRM without hiding any symbols.
Tests are now structered into `internal_api` and `public_api`. The
public API tests use the standard version of GRM with hidden internal
symbols.
For consistency reasons, the old test Makefile is replaced with a
CMakeLists file as well.