Eliminate integer cut offs in the Qt plugin
Many operations in the Qt plugin used integer rounding to position
values on the underlying pixmap grid. This behavior is correct when
logical pixels are physical pixels. However, for Retina or HiDPI
displays that is not the case, which leads to a effective resolution
reduction (by factor 2 in most cases). Therefore, this commit replaces
integer with floating point operations and removes integer cut offs. Qt
can now use fractional (logical) pixels and maps these to physical
pixels itself.