Change the `hist` to calculate the histogram
The histogram is expecting a list of `x` values which are then used to
calculate a histogram. Furthermore it is possible to set the number of
bins (`nbins` type: int) and the weights (`weights` type: list of double)
to change the output.

Colors can be specified with `bar_color` and `edge_color` as a subplot
argument. Either can be a single integer representing the gr colorindex
or a normalized rgb list.

Also added a histogram example.