[GRM] Handle `xlim`, `ylim`, `zlim` and `clim` as series attributes
Limits can now be given for each series. If no limits are passed by the
user, the `plot_store_coordinate_ranges` function will store the missing
limits so plotting routines can rely on the availability of the limits
attributes. The internal `xrange`, `yrange`, `zrange` and `crange`
attributes are now defined as the smallest rectangle which includes the
limits of all series. Thus, if no limits are given by the user, the
range attributes (and therefore the gr window) surely contain all series
data.