[GRM] Fix handling of character arrays (`C`)
The `C` (char array) format should be handled like `s` (string) in the
argument container to unify the input data as much as possible. However,
`C` was converted to `ns` instead (string with array length) which is an
invalid format (only `nS` is allowed) and led to program crashes. This
commit removes the `n` from the format string.