Avoid a name clash between Windows `io.h` and qhull `io.h`
Before this commit, the Windows header `io.h` was overlaid by `io.h`
from the qhull include directory. This commit removes the qhull include
directory from the include search list and renames the include directory
from `qhull` to `libqhull` in the 3rdparty directory, so qhull can be
included with `#include "libqhull/libqhull_a.h"`. The `libqhull` prefix
is the recommended way of the qhull authors to include the qhull header
files.