Add a function protocol to the socket driver
This commit adds a function or control byte which precedes the previous
socket driver protocol. The function bytes are taken from the Quartz
plugin and thus are compatible with each other. Using a function byte
enables more control over the running server and allows to query for
information (next commit).