Fix wrong resize calculations in gksqt under tiling window managers
In floating window managers, there is always a paint event before a
user-generated resize event. Thus, in floating windows managers the
first interpretation of the display list always initializes the wswindow
(`p->prevent_resize` is `0`). In tiling window managers, the first
user-generated resize event is queued before the first paint event (the
wm fits the window into a tile). In this case, wswindow would never be
set. Therefore, always initialize wswindow if this is the first
interpretation of a display list.