[GRM] Allow single scalar values to be read as arrays
All values except single scalar values are stored as array types in the
GRM argument container. Arrays can be read as arrays or as a series of
single scalars. However, single scalars could not be read as an
one-element array. This commit extends `args_first_value` and
`args_values` to allow single scalars to be read as array types. See
`args_automatic_array_conversion.c` for examples.