Fix bad polyline performance in the Qt plugin
Qt is slow on computing line joins for a large list of points.
Therefore, this commit adds a `MAX_POINTS_PERFORMANCE_THRESHOLD`
threshold constant. If this threshold is exceeded, the line width of
polylines is set to `0` which causes Qt to always draw a one pixel wide
line without the need to calculate any line joins.