[GRM] Fix public `layout.h` and `layout.hpp` headers
`layout.h` and `layout.hpp` are public headers (for library users) but
included the internal `args_int.h` header which is not installed on
`cmake --install` or `make install`. This commit fixes that by only
including the public `args.h` header and using `args_int.h` in the C and
C++ modules themselves.