[GRM] Add a request event
If a merged (or plotted) argument container contains a `request` key,
the container is supposed to be a request object. The corresponding
value must be a string which describes the request. The request is
enqueued as a request event which must be processed by an event handler
to answer the request (for example over the network by calling
`grm_send`).