[grplot] Implement an automatic `GRDIR` detection mechanism
If GR is used in another location than the original install location,
the hard coded `GRDIR` string points to a wrong directory. In this case,
the correct location must be set with the `GRDIR` environment variable.
This commit adds a mechanism which sets the correct `GRDIR` in `grplot`
automatically based on the executable path.