Improve `GRDIR` detection in `grplot`
Improve the `GRDIR` detection by checking the existence of the header
file `include/gr.h` in the found gr directory. This additional check
is necessary to detect if `grplot` was called from a regular GR
installation (like `/usr/local/gr`) or from the GR source or build
directory without installation.