[GRM] Revise `surface`, support missing `x` and / or `y` values
This commit adds support for z-only surface plots and for missing `x` or
`y` values. Missing values are substituted with equidistant values. If
neither `x` nor `y` are given, an additional `z_dims` variable must be
set for the exact shape of `z`. The implementation is very similar to
the already existing heatmap implementation.