[Qt plugin] Handle the `CLEARWS` code in `gks_dl_write_item`
Before this commit, a `CLEARWS` code caused the deletion of the whole
display list in the qt driver, before some elements of the display list
could be kept in a later call of the `gks_dl_write_item` routine.
Therefore, this commit removes the handling of `CLEARWS` in the Qt
plugin completely.