[GRM] Allow network address reuse in `net.c`
By default, address/port combinations which were bound to a TCP
connection and sent a FIN packet cannot be used to establish a new
connection in a specific time interval (as long as the port is in
`TIME_WAIT` state). This causes connections problems when `grplot` is
auto started, since the first connection to grplot is used to check its
alive status, preventing a following data connection from the GRM
library. Thus, this commit sets the `SO_REUSEADDR` flag on connections
established in GRM.