[Windows] Fix creation of grplot command line in `gr_startlistener`
If the `GRDIR` environment variable is not set, `gr_startlistener` falls
back to the hard coded `GRDIR` macro and puts it into the command line
template with a `%wS` format specifier which is no valid format
(probably wrongly derived from `%ws` which always accepts a wide
string). This commit replaces `%wS` and `%ws` with `%S` and `%s`.