[grplot] Improve network code for animations
Before this commit, receiving data and painting were not synchronized.
Multiple data frames could be received before a single image was drawn
on screen if the Qt event loop was busy. This commit adds a
synchronization mechanism. New data is only received after a previous
data frame was processed and displayed.