[grplot] Call `grm_plot` only on first draw (or always in listen mode)
The previous commits causes the graphics tree to always be recreated
from the argument container on `grm_plot` calls. Therefore, `grm_render`
must be called instead in `grplot`, to preserve the graphics tree.
However, `grplot` also depends on the container structure (for input for
example). Therefore, call `grm_plot` on the first draw operation, to
initialize the internal argument container.