[GRM] Dump backtraces on signal SEGV and ABRT if `GRM_BACKTRACE` is set
Dump backtraces on signal *Segmentation Fault* and *Abort* if the
environment variable `GRM_BACKTRACE` is set and contains `1`, `on`,
`true` or `yes` (also capitalized allowed). Backtraces are generated
with `backtrace_symbols_fd`, stored into a file `grm_backtrace` in the
system's temp directory. and need be translated into a human readable
format with `addr2line` afterwards.