[GRM] Avoid multiple render passes during one render step
Because of automatic tree updates, multiple render passes can occur
during a single render call. This is problematic since it decreases
performance and leads to bounding box IDs which are never freed again.
(IDs are reserved during the render call but only activated when the
image is finally drawn. Non-activated IDs are not freed anymore.)