Add support for the Kitty graphics protocol and send data in chunks
This commit adds a new workstation type `152` to generate escape codes
to encode a PNG image for the Kitty terminal (and compatible ones).
Kitty can be auto-detected and is then used by default if no graphical
display is available (mostly over SSH). Graphics data is sent in chunks
(also for iTerm now) to be more compatible with traditional terminal
escape codes which are limited in size.

In order to work in tmux, the tmux option `allow-passthrough` must be
activated for tmux version >= 3.3.