Fix broken XML import on macOS Tahoe
When processing XML files, internal attributes are decoded from their
BSON/Base64 representation on the fly. In order to check if delimiting
spaces between attributes must be inserted, `rdbuf()->in_avail()` is
used on a stringstream object to check if another attribute has already
been written to the stream. However, this is not reliable and can be
false even if data was written to the stream before. Thus, this commit
uses an additional flag to check if a delimiting space must be inserted
or not.