Synchronize close requests to gksqt
Before this commit, GKS did not wait for a response when sending a close
request to gksqt. This could lead to broken pipes when new connections
to a closing gksqt application were created fast enough. Now, GKS waits
for a response after sending a close request, giving gksqt time to close
its server socket to not accept any new connections.